dags/cf/get-data-upload-to-gcs/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-storage
pandas
fastparquet
gcsfs